==> notebooks/09_lightgbm_model_with_feature_engineering.ipynb <==
import pandas as pd

from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df.head(10)

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
  df, 
  datetime(2022, 6, 1, 0, 0, 0),
  target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

def average_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
  """
  Adds one column wtih the average rides from
  - 7 days ago
  - 14 days ago
  - 21 days ago
  - 28 days ago
  """
  X['average_rides_last_4_weeks'] = 0.25 * (
    X[f'rides_previous_{7*24}_hour'] + \
    X[f'rides_previous_{2*7*24}_hour'] + \
    X[f'rides_previous_{3*7*24}_hour'] + \
    X[f'rides_previous_{4*7*24}_hour'] 
  )
  return X

from sklearn.preprocessing import FunctionTransformer

add_feature_average_rides_last_4_weeks = FunctionTransformer(
  average_rides_last_4_weeks, validate=False
)

add_feature_average_rides_last_4_weeks.fit_transform(X_train)

from sklearn.base import BaseEstimator, TransformerMixin

class Temporal_Feature_Engineer(BaseEstimator, TransformerMixin):
  def fit(self, X, y=None):
    return self
  
  def transform(self, X, y=None):
    X_ = X.copy()

    # Generate numeric columns from datetime
    X_["hour"] = X_["pickup_hour"].dt.hour
    X_["day_of_week"] = X_["pickup_hour"].dt.dayofweek

    return X_.drop(columns=["pickup_hour"])

add_temporal_features = Temporal_Feature_Engineer()
add_temporal_features.fit_transform(X_train) 

import lightgbm as lgb

from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
  add_feature_average_rides_last_4_weeks,
  add_temporal_features,
  lgb.LGBMRegressor()
)
pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:0.4f}')
